test: consolidate AssessmentForm tests and add date range validation coverage
