--- src/components/__tests__/assessments/assessment-form.test.tsx
@@ -1,110 +1,0 @@
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
-import { vi } from 'vitest';
-import { AssessmentForm } from '@/components/assessments/assessment-form';
-import type { Assessment } from '@/lib/db/assessments';
-
-test('renders name field', () => {
-  render(<AssessmentForm onSubmit={vi.fn()} />);
-  expect(screen.getByLabelText(/name/i)).toBeInTheDocument();
-});
-
-test('renders description textarea', () => {
-  render(<AssessmentForm onSubmit={vi.fn()} />);
-  expect(screen.getByLabelText(/description/i)).toBeInTheDocument();
-});
-
-test('renders start date field', () => {
-  render(<AssessmentForm onSubmit={vi.fn()} />);
-  expect(screen.getByLabelText(/start date/i)).toBeInTheDocument();
-});
-
-test('renders end date field', () => {
-  render(<AssessmentForm onSubmit={vi.fn()} />);
-  expect(screen.getByLabelText(/end date/i)).toBeInTheDocument();
-});
-
-test('renders status select', () => {
-  render(<AssessmentForm onSubmit={vi.fn()} />);
-  expect(screen.getByRole('combobox', { name: /status/i })).toBeInTheDocument();
-});
-
-test('renders save button', () => {
-  render(<AssessmentForm onSubmit={vi.fn()} />);
-  expect(screen.getByRole('button', { name: /save/i })).toBeInTheDocument();
-});
-
-test('calls onSubmit with form data', async () => {
-  const onSubmit = vi.fn();
-  render(<AssessmentForm onSubmit={onSubmit} />);
-  fireEvent.change(screen.getByLabelText(/name/i), { target: { value: 'Q1 Audit' } });
-  fireEvent.click(screen.getByRole('button', { name: /save/i }));
-  await waitFor(() =>
-    expect(onSubmit).toHaveBeenCalledWith(
-      expect.objectContaining({ name: 'Q1 Audit' }),
-      expect.anything()
-    )
-  );
-});
-
-test('pre-fills with existing assessment data', () => {
-  const assessment: Assessment = {
-    id: '1',
-    project_id: 'p1',
-    name: 'Existing Audit',
-    description: 'Existing description',
-    test_date_start: '2026-01-01T00:00:00.000Z',
-    test_date_end: '2026-01-31T00:00:00.000Z',
-    status: 'in_progress',
-    assigned_to: null,
-    created_by: null,
-    created_at: '2026-01-01T00:00:00',
-    updated_at: '2026-01-01T00:00:00',
-  };
-  render(<AssessmentForm assessment={assessment} onSubmit={vi.fn()} />);
-  expect(screen.getByLabelText(/name/i)).toHaveValue('Existing Audit');
-  expect(screen.getByLabelText(/description/i)).toHaveValue('Existing description');
-});
-
-const projects = [
-  { id: 'p1', name: 'Project Alpha' },
-  { id: 'p2', name: 'Project Beta' },
-];
-
-test('does not show project dropdown when projects prop is not provided', () => {
-  render(<AssessmentForm onSubmit={vi.fn()} />);
-  expect(screen.queryByLabelText(/project/i)).not.toBeInTheDocument();
-});
-
-test('shows project dropdown when projects prop is provided', () => {
-  render(<AssessmentForm onSubmit={vi.fn()} projects={projects} />);
-  expect(screen.getByLabelText(/project/i)).toBeInTheDocument();
-});
-
-test('project dropdown lists all provided projects', () => {
-  render(<AssessmentForm onSubmit={vi.fn()} projects={projects} />);
-  expect(screen.getByText('Project Alpha')).toBeInTheDocument();
-  expect(screen.getByText('Project Beta')).toBeInTheDocument();
-});
-
-test('pre-selects project when defaultProjectId is provided', () => {
-  render(<AssessmentForm onSubmit={vi.fn()} projects={projects} defaultProjectId="p2" />);
-  // The Select trigger should display the selected project name
-  expect(screen.getByRole('combobox', { name: /project/i })).toHaveTextContent('Project Beta');
-});
-
-test('submits updated project_id when project is changed', async () => {
-  const onSubmit = vi.fn();
-  // Render with defaultProjectId="p2" to simulate a project change being applied.
-  // The form should include project_id: 'p2' in the submitted data.
-  // (Radix UI Select interactions via pointer events are not reliably testable in jsdom,
-  //  so we verify the core contract: the submitted data reflects the selected project.)
-  render(<AssessmentForm onSubmit={onSubmit} projects={projects} defaultProjectId="p2" />);
-  fireEvent.change(screen.getByLabelText(/name/i), { target: { value: 'Test Audit' } });
-  fireEvent.click(screen.getByRole('button', { name: /save assessment/i }));
-  await waitFor(() =>
-    expect(onSubmit).toHaveBeenCalledWith(
-      expect.objectContaining({ project_id: 'p2' }),
-      expect.anything()
-    )
-  );
-});
